refactor Added all tests to the appropriate places and cleared duplicate tests
