+++ app/src/util/tools/fpmm/liquidity/calcPoolTokens.ts
@@ -1,5 +1,5 @@
 import { BigNumber } from 'ethers/utils'
-//stays checkkkc
+
 /**
  * Compute the number of liquidity pool tokens that will be sent to the user by the Market Maker
  * after adding `addedFunds` of collateral.
